armsr: armv8: Set configs using arch/arm64/configs/defconfig
It's preferable the kernel configs come from the Linux kernel upstream
(arch/arm64/configs/defconfig), as they are likely to have all the
necessary configurations added by people with more expertise on the
particular board, as opposed to adding SystemReady-Certified boards one
at a time along with their drivers.

OpenWrt global generic and board configs are not required because
upstream becomes the starting point for the initial configuration. It's
only keeping the sub-target for required adjustments to be in sync with
OpenWrt kernel packages.

Signed-off-by: Javier Tia <javier.tia@linaro.org>