kernel: armsr: Allow to use a different init for initramfs
On ARM SystemReady platforms, it is required to have separate initramfs
to load kernel modules at boot time and before reaching the real rootfs.
The current generic init script mounts the rootfs in tmpfs, but it's
required to mount the real rootfs.

Waiting for the root files is required, as some storage takes time to be
detected by the kernel, especially the USB storage.

Signed-off-by: Javier Tia <javier.tia@linaro.org>