kernel: armsr: Load kernel modules with kmod from module-list.kmod
OpenWrt uses kmodloader to load the kernel modules, but it doesn't
support looking up the modules when they are installed by
install_modules. kmod solves it.

As kmod-* are not used anymore and some of them are still installed by
default, it's required to clean up the modules to be loaded by
kmodloader when it reads the /etc/modules* configurations. 

Signed-off-by: Javier Tia <javier.tia@linaro.org>