kernel: armsr: Use depmod host for module dependency data in initrd
When kernel modules are installed, the kernel build system fails
silently if depmod is not installed from the kmod package and module
dependency data is not generated. The kmod package is not installed by
default, like in Linux docker images, and kmod older versions are not
compatible with more recent kernels.

It's required to call host depmod again because a short list of kernel
modules is installed in initrd. Passing the '-e' flag will detect a
missing module dependency in case more modules are added.

Build kmod host only on armsr target.

Signed-off-by: Javier Tia <javier.tia@linaro.org>