armsr: Autoprobe kernel modules with Busybox mdev
A picked kernel module list has been added in initrd. Not all those
modules are required to be loaded; it depends on the HW. It's preferred
autoprobe modules when certain HWs have been identified by the kernel.
udev supports autoprobe kernel modules, but it's not supported by
OpenWrt.

In OpenWrt, procd is the default device manager, and hotplug is
supported, but there is no indication in the code or Wiki documentation
that autoprobe is supported and it runs after the initrd has been
loaded.

Busybox has mdev as its device manager, which can support autoprobe with
autoplug.

Signed-off-by: Javier Tia <javier.tia@linaro.org>