armsr: Compress kernel modules with zstd
The kernel modules are compressed with gzip and initrd with xz. Loading
initrd with all kernel modules increases the boot time to ~10s because
of the slowness of gzip/xz compressing and decompressing. With zstd, the
boot time is decreased to ~5s from ~10s with gzip/xz. zstd provides the
best compromise between compression ratio and time. Even though xz
produces the shortest module size, the compression process takes a lot
longer.

Signed-off-by: Javier Tia <javier.tia@linaro.org>